move people_matcher views to people_admin
