+++ dashboards/urls.py
@@ -5,9 +5,6 @@ from .views import (
     dq_overview,
     dq_overview_session,
     dqr_listing,
-    people_list,
-    people_matcher,
-    people_matcher_session,
 )
 from utils.common import states
 
@@ -23,15 +20,5 @@ urlpatterns = [
     re_path(
         r"^dq_overview/(?P<state>{})/(?P<session>[-\w ]+)/$".format(state_abbr_pattern),
         dq_overview_session,
-    ),
-    path("people/", people_list),
-    re_path(
-        r"^people/(?P<state>{})_matcher/$".format(state_abbr_pattern), people_matcher
-    ),
-    re_path(
-        r"^people/(?P<state>{})_matcher/(?P<session>[-\w ]+)/$".format(
-            state_abbr_pattern
-        ),
-        people_matcher_session,
     ),
 ]
