+++ web/urls.py
@@ -10,6 +10,7 @@ from bundles.views import bundle_view
 
 urlpatterns = [
     path("djadmin/", admin.site.urls),
+    path("admin/people/", include("people_admin.urls")),
     path("accounts/", include("allauth.urls")),
     path("accounts/profile/", include("profiles.urls")),
     path("dashboard/", include("dashboards.urls")),
